+++ ionoscloud/test_constants.go
@@ -1852,6 +1852,9 @@ const clusterConnectionsCidrAttribute = "cidr"
 const clusterDisplayNameAttribute = "display_name"
 const clusterMaintenanceWindowAttribute = "maintenance_window"
 const clusterMaintenanceWindowDayOfTheWeekAttribute = "day_of_the_week"
+const clusterBackupAttribute = "backup"
+const clusterBackupLocationAttribute = "location"
+
 const clusterMaintenanceWindowTimeAttribute = "time"
 const clusterCredentialsAttribute = "credentials"
 const clusterCredentialsUsernameAttribute = "username"
@@ -1866,6 +1869,8 @@ const clusterMaintenanceWindowDayOfTheWeekValue = "Sunday"
 const clusterMaintenanceWindowDayOfTheWeekUpdateValue = "Monday"
 const clusterMaintenanceWindowTimeValue = "09:00:00"
 const clusterMaintenanceWindowTimeUpdateValue = "10:00:00"
+
+const clusterBackupLocationValue = "de"
 
 const wireguardGatewayConfig = `
 resource "ionoscloud_datacenter" "datacenter_example" {